backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-simple-history
gunicorn
psycopg2-binary
python-dotenv
celery>=5.3.6
Django>=4.2
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
django-guardian
openai
